fix outline and add useIsMobile hook
