+++ packages/react/src/hooks/useIsMobile.ts
@@ -1,0 +1,28 @@
+import { useState, useEffect } from 'react';
+
+const useIsMobile = () => {
+  const breakpoint = 768;
+  const [isMobile, setIsMobile] = useState(false);
+
+  const handleResize = () => {
+    if (window.innerWidth < breakpoint) {
+      setIsMobile(true);
+    } else {
+      setIsMobile(false);
+    }
+  };
+
+  useEffect(() => {
+    handleResize();
+
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+  return isMobile;
+};
+
+export default useIsMobile;
